runtime initialization done (debugged)

+++ arch/riscv/drivers/riscv64.h
@@ -62,12 +62,12 @@ static inline void flush_tlb(void)
   asm volatile ("sfence.vma");
 }
 
-static inline pte_t pte_create(uintptr_t ppn, int type)
+static inline pte_t pte_create(unsigned long ppn, int type)
 { 
   return __pte( (ppn << PTE_PPN_SHIFT) | PTE_V | type );
 }
 
-static inline pte_t ptd_create(uintptr_t ppn)
+static inline pte_t ptd_create(unsigned long ppn)
 {
   return pte_create(ppn, PTE_V);
 }
